add html,js edit EntityController,ScheduleRepository,ScheduleService,ScheduleServiceImpl

+++ src/main/java/com/university/university/repository/LessonRepository.java
@@ -3,5 +3,5 @@ package com.university.university.repository;
 import com.university.university.entity.Lesson;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LessonRepository extends JpaRepository<Lesson, Integer> {
+public interface LessonRepository extends JpaRepository<Lesson, Long> {
 }
