+++ src/main/java/com/university/university/repository/TeacherRepository.java
@@ -2,6 +2,6 @@ package com.university.university.repository;
 
 import com.university.university.entity.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Integer> {}
+public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+}
